# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: SkillScan API
  version: "1.1"
  description: |
    SkillScan 团队使用的统一接口。
    包含健康检查、GitHub 安装绑定、报告上传、分享页、扫描记录、审批和审计。
servers:
  - url: https://api.skillscan.example.com
    description: Replace with your real base URL
paths:
...
  /github/webhooks:
    post:
      summary: GitHub App webhook
      security: []
      parameters:
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema:
            type: string
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processed
        "401":
          description: Invalid signature
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...